Setup Axios Interceptors

+++ src/actions/auth.actions.js
@@ -61,10 +61,23 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
 	return async (dispatch) => {
-		// TODO Logout API Call
-		localStorage.clear();
 		dispatch({
 			type: authConstants.LOGOUT_REQUEST,
 		});
+		const res = await axios.get("/users/auth/signout");
+
+		if (res.status === 200) {
+			localStorage.clear();
+			dispatch({
+				type: authConstants.LOGOUT_SUCCESS,
+			});
+		} else {
+			dispatch({
+				type: authConstants.LOGOUT_FAILURE,
+				payload: {
+					error: res.message,
+				},
+			});
+		}
 	};
 };
